requirements.txt
requests
selenium
bs4
urllib3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas